+++ test/mocha.opts
@@ -1,1 +1,2 @@
 --exit
+--recursive
