+++ test/redis/processReminders.js
@@ -1,20 +1,20 @@
 /* global describe beforeEach context it */
 
-require('./helper.js');
+require('./../helper.js');
 
 const proxyquire = require('proxyquire').noCallThru();
 const { expect } = require('chai');
 const moment     = require('moment');
-const msg        = require('./stubs/message.js');
-const tzStore    = require('../api/redis/client.js').timezones;
-const reminders  = require('../api/redis/client.js').reminders;
+const msg        = require('./../stubs/message.js');
+const tzStore    = require('../../api/redis/client.js').timezones;
+const reminders  = require('../../api/redis/client.js').reminders;
 
 /* eslint-disable global-require */
 const subject = proxyquire(
-  '../api/redis/processReminders.js',
+  '../../api/redis/processReminders.js',
   {
     // 'discord.js-commando': require('./stubs/Command.js'),
-    './../bot.js': new (require('./stubs/CommandoClient.js').CommandoClient)()
+    './../bot.js': new (require('./../stubs/CommandoClient.js').CommandoClient)()
   }
 );
 /* eslint-enable global-require */
